Resolve AGM route casing conflict in Next build
